Move FirebaseAuthManager to its own file
